app-v1/requirements.txt
chainlit
langchain
langchain-community
langchain_openai
